Storefront
The bracelets are a light purple color to indicate eating disorder awareness. They come in both large & small sizes. Price will be reduced if you are buying in bulk. Contact me for ordering and details...Be sure to follow some of the links on my page - I find them to be greatly beneficial.